tech_sentiment/requirements.txt
alpaca-trade-api
pandas
numpy
ta-lib
requests
python-dotenv
beautifulsoup4
nltk
finviz
bs4
